@@ -2,6 +2,7 @@
 #define WEBCC_COMMON_H_
 
 #include <cassert>
+#include <filesystem>
 #include <string>
 #include <utility>
 #include <vector>
@@ -156,7 +157,8 @@ public:
   // Construct a file part.
   // The file name will be extracted from path.
   // The media type, if not provided, will be inferred from file extension.
-  static FormPartPtr NewFile(const std::string& name, const Path& path,
+  static FormPartPtr NewFile(const std::string& name,
+                             const std::filesystem::path& path,
                              const std::string& media_type = "");
 
   // API: SERVER
@@ -227,7 +229,7 @@ private:
   std::string name_;
 
   // The path of the file to post.
-  Path path_;
+  std::filesystem::path path_;
 
   // The original local file name.
   // E.g., "baby.jpg".
