@@ -1,17 +1,15 @@
 #include "webcc/parser.h"
 
 #include "boost/algorithm/string.hpp"
-#include "boost/filesystem/operations.hpp"
 
 #include "webcc/logger.h"
 #include "webcc/message.h"
+#include "webcc/string.h"
 #include "webcc/utility.h"
 
 #if WEBCC_ENABLE_GZIP
 #include "webcc/gzip.h"
 #endif
-
-namespace bfs = boost::filesystem;
 
 namespace webcc {
 
@@ -69,11 +67,17 @@ bool StringBodyHandler::Finish() {
 
 bool FileBodyHandler::OpenFile() {
   try {
-    temp_path_ = bfs::temp_directory_path() / bfs::unique_path();
+    temp_path_ = std::filesystem::temp_directory_path();
+
+    // Generate a random string as file name.
+    // A replacement of boost::filesystem::unique_path().
+    temp_path_ /= string::RandomString(10);
+
     LOG_VERB("Generate a temp path for streaming: %s",
              temp_path_.string().c_str());
-  } catch (const bfs::filesystem_error&) {
-    LOG_ERRO("Failed to generate temp path: %s", temp_path_.string().c_str());
+
+  } catch (const std::filesystem::filesystem_error&) {
+    LOG_ERRO("Failed to generate temp path for streaming.");
     return false;
   }
 
