@@ -5,6 +5,13 @@
 
 #include "webcc/globals.h"
 
+// Avoid include <filesystem> in the header.
+namespace std {
+namespace filesystem {
+class path;
+}  // namespace filesystem
+}  // namespace std
+
 namespace webcc {
 namespace utility {
 
@@ -29,10 +36,10 @@ bool ToSize(const std::string& str, int base, std::size_t* size);
 
 // Tell the size in bytes of the given file.
 // Return kInvalidLength (-1) on failure.
-std::size_t TellSize(const Path& path);
+std::size_t TellSize(const std::filesystem::path& path);
 
 // Read entire file into string.
-bool ReadFile(const Path& path, std::string* output);
+bool ReadFile(const std::filesystem::path& path, std::string* output);
 
 // Dump the string data line by line to achieve more readability.
 // Also limit the maximum size of the data to be dumped.
